deleting old components and moving rest to registrations folder
