+++ src/components/Registrations/Registrations.vue
@@ -28,27 +28,14 @@
       title="Afmelden van de les"
       ok-only ok-title="Niet afmelden"
     >
-      <Absence :registration="absenceModal.registration" @hide="resetAbsenceModal"/>
+      <AbsenceForm :registration="absenceModal.registration" @hide="resetAbsenceModal"/>
     </b-modal>
-    <b-table-simple striped>
-      <b-thead>
-        <b-tr>
-          <b-th>Naam</b-th>
-          <b-th>Studie</b-th>
-          <b-th>Klas</b-th>
-        </b-tr>
-      </b-thead>
-      <b-tbody>
-        <Registration ref="registration" v-for="registration in registrations" v-bind:key="registration.id" v-bind:registration="registration"/>
-      </b-tbody>
-    </b-table-simple>
   </b-container>
 </template>
 
 <script>
   import Vue from "vue";
-  import Registration from "../Registration/Registration.vue";
-  import Absence from "../Absence/Absence.vue";
+  import AbsenceForm from "./AbsenceForm.vue";
   import StudentLessons from "./StudentLessons.vue";
   import { EventBus } from "../../EventBus";
   import { deleteStudent} from "@/firebase/firebase.js";
@@ -56,8 +43,7 @@
   export default Vue.extend ({
     name: "Registrations",
     components: {
-      Registration,
-      Absence,
+      AbsenceForm,
       StudentLessons
     },
     props: ["registrations"],
