+++ src/views/Home.vue
@@ -1,7 +1,6 @@
 <template>
   <b-container>
-    <Registrations v-bind:registrations="registrations"/>
-    <Attendance />
+    <Registrations :registrations="registrations"/>
     <br><b-button v-b-toggle.collapse-registration variant="primary"> Schrijf uzelf of uw kind in!</b-button><br>
     <b-collapse id="collapse-registration" class="mt-2">
       <RegistrationForm />
@@ -11,7 +10,6 @@
 
 <script>
   import Vue from "vue";
-  import Attendance from "../components/Attendance/Attendance.vue";
   import RegistrationForm from "../components/RegistrationForm/RegistrationForm.vue";
   import Registrations from "../components/Registrations/Registrations.vue";
   import { EventBus } from "../EventBus";
@@ -21,7 +19,6 @@
   export default Vue.extend({
     name: "Home",
     components: {
-      Attendance,
       RegistrationForm,
       Registrations
     },
